Update and rename config.js to Suhail-Md/config.js
update and rename config.js to sahail-Md/config.js
